DYL/requirements.txt
asgiref==3.2.10
astroid==2.4.2
boto3==1.17.30
botocore==1.20.30
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.1
django-import-export==2.3.0
django-request==1.5.6
django-storages==1.11.1
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
isort==5.6.3
jdcal==1.4.1
jmespath==0.10.0
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.5
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.3.4
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
